requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.17

mysqlclient==2.1.1
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
